improved key controls system applied to rest of existing logic, todo much later button reassignment

+++ source/common_utils.h
@@ -46,7 +46,7 @@ typedef u64 usize;
 #define gb(n) (mb(n) * 1024ull)
 
 
-#define foreach(i, lim) for (u64 (i) = 0; (i) < (lim); ++(i))
+#define foreach(i, lim) for (u64 (i) = 0; (i) < ((u64)lim); ++(i))
 #define forrange(i, l, h) for (i64 (i) = (l); (i) < (h); ++(i))
 
 // }
